+++ src/main/java/agents/markets/meritOrder/books/OrderBookItem.java
@@ -4,7 +4,6 @@
 package agents.markets.meritOrder.books;
 
 import java.util.Comparator;
-
 import agents.markets.meritOrder.Bid;
 import agents.markets.meritOrder.Bid.Type;
 import de.dlr.gitlab.fame.communication.transfer.ComponentCollector;
@@ -13,7 +12,7 @@ import de.dlr.gitlab.fame.communication.transfer.Portable;
 
 /** An item of an {@link OrderBook}, either bid or ask
  * 
- * @author Christoph Schimeczek, Martin Klein , Evelyn Sperber, Farzad Sarfarazi */
+ * @author Christoph Schimeczek, Martin Klein , Evelyn Sperber, Farzad Sarfarazi, A. Achraf El Ghazi */
 public class OrderBookItem implements Portable {
 	static final String ERR_NEGATIVE_POWER = "OrderBookItems with negative power received from Trader: ";
 
@@ -35,18 +34,9 @@ public class OrderBookItem implements Portable {
 		}
 	}
 
-	/** Sets the cumulated power upper value in the context of the containing {@link OrderBook}
-	 * 
-	 * @param cumulatedPower sum of all previous bids in the merit-order plus the amount of power in this item's {@link Bid} */
-	void setCumulatedPowerUpperValue(double cumulatedPower) {
-		this.cumulatedPowerUpperValue = cumulatedPower;
-	}
-
-	/** Sets the actual awarded power in the context of the containing {@link OrderBook}
-	 * 
-	 * @param awardedPower of this bid; any value between 0 and {@link #getBlockPower()} */
-	void setAwardedPower(double awardedPower) {
-		this.awardedPower = awardedPower;
+	/** @return the bid of the OrderBookItem */
+	public Bid getBid() {
+		return bid;
 	}
 
 	/** @return actual awarded power; call only after market clearing */
@@ -59,9 +49,23 @@ public class OrderBookItem implements Portable {
 		return getBlockPower() - awardedPower;
 	}
 
+	/** Sets the actual awarded power in the context of the containing {@link OrderBook}
+	 * 
+	 * @param awardedPower of this bid; any value between 0 and {@link #getBlockPower()} */
+	public void setAwardedPower(double awardedPower) {
+		this.awardedPower = awardedPower;
+	}
+
 	/** @return sum of all previous bids in the merit-order <b>plus</b> the amount of power in this item's {@link Bid} */
 	public double getCumulatedPowerUpperValue() {
 		return cumulatedPowerUpperValue;
+	}
+
+	/** Sets the cumulated power upper value in the context of the containing {@link OrderBook}
+	 * 
+	 * @param cumulatedPower sum of all previous bids in the merit-order plus the amount of power in this item's {@link Bid} */
+	public void setCumulatedPowerUpperValue(double cumulatedPower) {
+		this.cumulatedPowerUpperValue = cumulatedPower;
 	}
 
 	/** @return sum of all previous bids in the merit-order <b>without</b> the amount of power in this item's {@link Bid} */
@@ -84,25 +88,25 @@ public class OrderBookItem implements Portable {
 		return bid.getMarginalCost();
 	}
 
+	@Override
+	public String toString() {
+		return "Awarded: " + awardedPower + " " + bid;
+	}
+
 	/** @return ID of the traded that issued the associated {@link Bid} */
 	public long getTraderUuid() {
 		return bid.getTraderUuid();
 	}
 
 	@Override
-	public String toString() {
-		return "Awarded: " + awardedPower + " " + bid;
-	}
-
 	/** required for {@link Portable}s */
-	@Override
 	public void addComponentsTo(ComponentCollector collector) {
 		collector.storeComponents(bid);
 		collector.storeDoubles(cumulatedPowerUpperValue, awardedPower);
 	}
 
-	/** required for {@link Portable}s */
 	@Override
+	/** required for {@link Portable}s */
 	public void populate(ComponentProvider provider) {
 		bid = provider.nextComponent(Bid.class);
 		cumulatedPowerUpperValue = provider.nextDouble();